refactor: 점수(카드 숫자의 합)을 포장한 객체 구현
